fixup! [PAL/Linux-SGX] Use `clock_gettime()` from untrusted vDSO
