+++ pal/src/host/linux-sgx/host_ocalls.c
@@ -901,92 +901,3 @@ int start_rpc(size_t threads_cnt) {
 
     return 0;
 }
-
-bool g_gettime_helper_started = false;
-
-static int gettime_helper_loop(void* arg) {
-    int ret;
-    uint64_t* time_addr = (uint64_t*)arg;
-
-    /* block all signals for this helper thread */
-    __sigset_t mask;
-    __sigfillset(&mask);
-    DO_SYSCALL(rt_sigprocmask, SIG_SETMASK, &mask, NULL, sizeof(mask));
-
-    __atomic_store_n(&g_gettime_helper_started, true, __ATOMIC_RELAXED);
-
-    while (1) {
-        struct timespec time;
-        if (g_pal_enclave.vdso_clock_gettime) {
-            ret = g_pal_enclave.vdso_clock_gettime(CLOCK_REALTIME, &time);
-        } else {
-            ret = DO_SYSCALL(clock_gettime, CLOCK_REALTIME, &time);
-        }
-        if (ret < 0)
-            BUG();
-
-        uint64_t time_usec = 1000000 * (uint64_t)time.tv_sec + time.tv_nsec / 1000;
-        __atomic_store_n(time_addr, time_usec, __ATOMIC_RELEASE);
-
-        /* NOTE: This helper thread sleeps for 100ms, which limits the timestamp resolution. This
-         *       duration was chosen to balance timestamp resolution and CPU load. */
-        struct timespec ts = {.tv_sec = 0, .tv_nsec = 100UL * 1000 * 1000};
-        (void)DO_SYSCALL(nanosleep, &ts, /*rem=*/NULL);
-    }
-
-    /* NOTREACHED */
-    return 0;
-}
-
-int start_gettime_helper(uint64_t** out_time_addr) {
-    int ret;
-
-    uint64_t* time_addr = NULL;
-    void* stack = NULL;
-
-    time_addr = (uint64_t*)DO_SYSCALL(mmap, NULL, PRESET_PAGESIZE, PROT_READ | PROT_WRITE,
-                                      MAP_ANONYMOUS | MAP_PRIVATE, -1, 0);
-    if (IS_PTR_ERR(time_addr)) {
-        ret = PTR_TO_ERR(time_addr);
-        goto fail;
-    }
-
-    stack = (void*)DO_SYSCALL(mmap, NULL, PRESET_PAGESIZE * 2, PROT_READ | PROT_WRITE,
-                              MAP_PRIVATE | MAP_ANONYMOUS, -1, 0);
-    if (IS_PTR_ERR(stack)) {
-        ret = PTR_TO_ERR(stack);
-        goto fail;
-    }
-
-    void* stack_top = stack + PRESET_PAGESIZE * 2;
-    stack_top = ALIGN_DOWN_PTR(stack_top, 16);
-
-    /* NOTE: Instead of spawning the helper thread that periodically calls into the untrusted-world
-     *       vDSO library, we could copy the vDSO code into the enclave, figure out how this code
-     *       must access vvar memory region (which resides in untrusted mem) and call into this code
-     *       from within the enclave. However, vDSO/vvar interactions are internal Linux details and
-     *       don't have stable APIs. */
-    int dummy_parent_tid_field = 0;
-    ret = clone(gettime_helper_loop, stack_top,
-                CLONE_VM | CLONE_FS | CLONE_FILES | CLONE_SYSVSEM |
-                 CLONE_THREAD | CLONE_SIGHAND | CLONE_PTRACE | CLONE_PARENT_SETTID,
-                /*arg=*/time_addr, &dummy_parent_tid_field, /*tls=*/NULL, /*child_tid=*/NULL,
-                /*dummy*/thread_exit);
-    if (ret < 0) {
-        ret = -EPERM;
-        goto fail;
-    }
-
-    while (!__atomic_load_n(&g_gettime_helper_started, __ATOMIC_RELAXED))
-        CPU_RELAX();
-
-    *out_time_addr = time_addr;
-    return 0;
-
-fail:
-    if (time_addr)
-        DO_SYSCALL(munmap, time_addr, PRESET_PAGESIZE);
-    if (stack)
-        DO_SYSCALL(munmap, stack, PRESET_PAGESIZE * 2);
-    return ret;
-}
